reorg xpath1 tests into themed subtests
